Se suben correcciones en cuanto a tipo de datos y rangos numericos

+++ src/get_numbers.h
@@ -1,2 +1,2 @@
-int getIntValidacion(char mensajeIngreso[], char errorIngreso[], int maximo, int minimo);
-int getInt(char mensajeIngreso[]);
+long long int getIntValidacion(char mensajeIngreso[], char errorIngreso[], int maximo, int minimo);
+long long int getInt(char mensajeIngreso[]);
